+++ pkg/webui/console/containers/application-payload-formatters/uplink.js
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2021 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -42,8 +42,9 @@ import {
 import style from './application-payload-formatters.styl'
 
 const m = defineMessages({
+  title: 'Default uplink payload formatter',
   infoText:
-    'These payload formatters are executed on uplink messages from all end devices in this application. Note: end device level payload formatters have precedence.',
+    'You can use the "Payload formatter" tab of individual end devices to test uplink payload formatters and to define individual payload formatter settings per end device.',
 })
 @connect(
   state => {
@@ -73,6 +74,16 @@ class ApplicationPayloadFormatters extends React.PureComponent {
     updateLinkSuccess: PropTypes.func.isRequired,
   }
 
+  constructor(props) {
+    super(props)
+
+    const { formatters } = props
+
+    this.state = {
+      type: formatters.up_formatter || PAYLOAD_FORMATTER_TYPES.NONE,
+    }
+  }
+
   @bind
   async onSubmit(values) {
     const { appId, formatters } = this.props
@@ -98,22 +109,30 @@ class ApplicationPayloadFormatters extends React.PureComponent {
     updateLinkSuccess(link)
   }
 
+  @bind
+  onTypeChange(type) {
+    this.setState({ type })
+  }
+
   render() {
     const { formatters } = this.props
-    const applicationFormatterInfo = (
-      <Notification className={style.notification} small info content={m.infoText} />
-    )
+    const { type } = this.state
+
+    const isNoneType = type === PAYLOAD_FORMATTER_TYPES.NONE
 
     return (
       <React.Fragment>
-        <PageTitle title={sharedMessages.payloadFormattersUplink} />
-        {applicationFormatterInfo}
+        <PageTitle title={m.title} />
+        {!isNoneType && (
+          <Notification className={style.notification} small info content={m.infoText} />
+        )}
         <PayloadFormattersForm
           uplink
           onSubmit={this.onSubmit}
           onSubmitSuccess={this.onSubmitSuccess}
           initialType={formatters.up_formatter || PAYLOAD_FORMATTER_TYPES.NONE}
           initialParameter={formatters.up_formatter_parameter || ''}
+          onTypeChange={this.onTypeChange}
         />
       </React.Fragment>
     )